most of backend
added functions: get searches from google trend api, finance api for the ticker, gets the moving average for the trend api, applies the trend to the stock price to predicate.

need to work on a bit of the math and also the google trend api only has data from 01 to 15 so i used that to predict 1/22/2021 stock price. It actually works pretty well. 
Need to workout functionality with the whole project and bit of statistics.